Refactoring to move each collection into specific file.
