+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/HappyHours.kt
@@ -2,8 +2,11 @@ package com.jordangellatly.hermosahappyhour.ui.home
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -21,6 +24,34 @@ import com.jordangellatly.hermosahappyhour.ui.components.HermosaHappyHourSurface
 import com.jordangellatly.hermosahappyhour.ui.components.RestaurantImage
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 import java.util.*
+
+@Composable
+fun HappyHourCollection(
+    restaurants: List<Restaurant>,
+    onRestaurantClick: (Long) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    LazyRow(
+        modifier = modifier,
+        contentPadding = PaddingValues(start = 12.dp, end = 12.dp)
+    ) {
+        items(restaurants) { snack ->
+            HappyHourItem(snack, onRestaurantClick)
+        }
+    }
+}
+
+@Preview
+@Composable
+fun HappyHourCollectionPreview() {
+    HermosaHappyHourTheme {
+        val restaurants = sampleSearchRestaurantData
+        HappyHourCollection(
+            restaurants = restaurants,
+            onRestaurantClick = {}
+        )
+    }
+}
 
 @Composable
 fun HappyHourItem(
